IC4C-214: Changed RCI's get_string() arguments type to avoid casting.
	- Changed from size_t which depends on the architecture to uint32_t for string length.
	- Due to protocol, the length is always specified in an uint32_t.

+++ private/rci_binary_support.h
@@ -286,7 +286,7 @@ typedef struct
         } element;
 
         connector_element_value_t value;
-        size_t string_value_length;
+        uint32_t string_value_length;
 
         connector_remote_config_t callback_data;
     } shared;
